feat: adding the role feature to the app and separating the admin routes to the user route

+++ routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Enums\UserRoleEnum;
 use App\Http\Controllers\Auth\GenerateTokenController;
 use App\Http\Controllers\Url\CreateUrlController;
 use App\Http\Controllers\Url\DeleteUrlController;
@@ -17,10 +18,10 @@ Route::post('/login', GenerateTokenController::class);
 
 Route::post('/users', CreateUserController::class);
 Route::prefix('users')->middleware('auth:sanctum')->group(function (): void {
-    Route::get('/{id}', DetailUserController::class)->whereUuid('id');
-    Route::get('/', ListUsersController::class);
-    Route::patch('/{id}', UpdateUserController::class)->whereUuid('id');
-    Route::delete('/{id}', DeleteUserController::class)->whereUuid('id');
+    Route::get('/{id}', DetailUserController::class)->whereUuid('id')->middleware('role:'.UserRoleEnum::ADMIN->value);
+    Route::get('/', ListUsersController::class)->middleware('role:'.UserRoleEnum::ADMIN->value);
+    Route::patch('/{id}', UpdateUserController::class)->whereUuid('id')->middleware('role:'.UserRoleEnum::ADMIN->value);
+    Route::delete('/{id}', DeleteUserController::class)->whereUuid('id')->middleware('role:'.UserRoleEnum::ADMIN->value);
 });
 
 Route::prefix('urls')->middleware('auth:sanctum')->group(function (): void {
